docker added, tracking added
docker added, tracking added

docker added, tracking added

readme edited, github workflows fixed, price -> floatfield

black fixes

github workflows fixed

coverage from dev to main deps

tests fixed

minor fixes

db -> localhost

entrypoint added

testing method edited

+++ price/models.py
@@ -10,7 +10,7 @@ class Symbol(models.Model):
 
 class Price(models.Model):
     symbol = models.ForeignKey(Symbol, on_delete=models.CASCADE)
-    price = models.IntegerField(default=0)
+    price = models.FloatField(default=0.0)
 
     created_at = models.DateTimeField(auto_now_add=True)
